Integrate with systemd's listenfd protocol
This will be useful for servers in the future, but is already useful
during development when using the systemfd executable by allowing you to
already run the ssh client while the server is still compiling. Systemfd
keeps the port open and as soon as the server has started it will accept
the pending connection attempt.